Interaction System
so for anyone using the interaction system heres the rundown:

Put the "Interacting" script on the player. This has a constant 3m raycast in front of the player. If it collides with an object with the tag "Interactable" and the player presses the right mouse button, the object is deleted.

However, if the object has the TextTrigger script on it, instead the object will disable it's mesh renderer and box collider components and preform the Load function, which calls a coroutine that plays the text on screen.

FYI: I changed the tag from deletable to Interactable, i just thought it made more sense.

TLDR:

1. put Interacting script on player
2. tag interactable objects with the Interactable tag
3. add TextTrigger script to interactable objects
4. Set the text you want to play in the public field of the TextTrigger script